+++ rsciio/blockfile/_api.py
@@ -21,11 +21,9 @@ import logging
 import os
 import warnings
 
-import dask
 import dateutil
 import numpy as np
-from dask.diagnostics import ProgressBar
-from skimage import dtype_limits
+import skimage
 
 from rsciio._docstrings import (
     CHUNKS_READ_DOC,
@@ -36,19 +34,15 @@ from rsciio._docstrings import (
     SHOW_PROGRESSBAR_DOC,
     SIGNAL_DOC,
 )
-from rsciio.utils.date_time_tools import (
+from rsciio.utils import file
+from rsciio.utils._array import dict2sarray, sarray2dict
+from rsciio.utils._context_manager import get_progress_bar_context_manager
+from rsciio.utils._date_time import (
     datetime_to_serial_date,
     serial_date_to_ISO_format,
 )
-from rsciio.utils.distributed import memmap_distributed
-from rsciio.utils.skimage_exposure import rescale_intensity
-from rsciio.utils.tools import (
-    DTBox,
-    convert_units,
-    dict2sarray,
-    dummy_context_manager,
-    sarray2dict,
-)
+from rsciio.utils._dictionary import DTBox
+from rsciio.utils._skimage_exposure import rescale_intensity
 
 _logger = logging.getLogger(__name__)
 
@@ -117,6 +111,8 @@ def get_default_header(endianess="<"):
 
 
 def get_header_from_signal(signal, endianess="<"):
+    from rsciio.utils._units import convert_units
+
     header = get_default_header(endianess)
     if "blockfile_header" in signal["original_metadata"]:
         header = dict2sarray(
@@ -302,7 +298,7 @@ def file_reader(filename, lazy=False, chunks="auto", endianess="<"):
                 ("data", np.dtype(endianess + "u1"), signal_shape),
             ]
         )
-        data = memmap_distributed(
+        data = file.memmap_distributed(
             filename,
             chunks=chunks,
             offset=offset2,
@@ -413,13 +409,15 @@ def file_writer(
                 UserWarning,
             )
     elif intensity_scaling == "dtype":
-        original_scale = dtype_limits(signal["data"])
+        original_scale = skimage.dtype_limits(signal["data"])
         if original_scale[1] == 1.0:
             raise ValueError("Signals with float dtype can not use 'dtype'")
     elif intensity_scaling == "minmax":
         minimum = signal["data"].min()
         maximum = signal["data"].max()
         if signal["attributes"]["_lazy"]:
+            import dask
+
             minimum, maximum = dask.compute(minimum, maximum)
         original_scale = (minimum, maximum)
     elif intensity_scaling == "crop":
@@ -496,8 +494,7 @@ def file_writer(
     file_memmap["MAGIC"] = magics
     file_memmap["ID"] = ids
     if signal["attributes"]["_lazy"]:
-        cm = ProgressBar if show_progressbar else dummy_context_manager
-        with cm():
+        with get_progress_bar_context_manager(show_progressbar)():
             array_data.store(file_memmap["IMG"])
     else:
         file_memmap["IMG"] = array_data