+++ rsciio/tiff/_api.py
@@ -35,8 +35,8 @@ from rsciio._docstrings import (
     RETURNS_DOC,
     SIGNAL_DOC,
 )
-from rsciio.utils.date_time_tools import get_date_time_from_metadata
-from rsciio.utils.tools import _UREG, DTBox
+from rsciio.utils._date_time import get_date_time_from_metadata
+from rsciio.utils._dictionary import DTBox
 
 _logger = logging.getLogger(__name__)
 
@@ -97,13 +97,13 @@ def file_writer(filename, signal, export_scale=True, extratags=None, **kwds):
     metadata = signal.get("metadata", {})
     photometric = "MINISBLACK"
     # HyperSpy uses struct arrays to store RGBA data
-    from rsciio.utils import rgb_tools
+    from rsciio.utils import rgb
 
     if extratags is None:
         extratags = []
 
-    if rgb_tools.is_rgbx(data):
-        data = rgb_tools.rgbx2regular_array(data)
+    if rgb.is_rgbx(data):
+        data = rgb.rgbx2regular_array(data)
         photometric = "RGB"
 
     if "description" in kwds.keys() and export_scale:
@@ -409,9 +409,9 @@ def _load_data(handle, is_rgb, memmap=None, **kwds):
     dc = handle.asarray(out=memmap)
     _logger.debug("data shape: {0}".format(dc.shape))
     if is_rgb:
-        from rsciio.utils import rgb_tools
+        from rsciio.utils import rgb
 
-        dc = rgb_tools.regular_array2rgbx(dc)
+        dc = rgb.regular_array2rgbx(dc)
 
     return dc
 
@@ -1099,6 +1099,8 @@ def _parse_tuple_Zeiss(tup):
 def _parse_tuple_Zeiss_with_units(tup, to_units=None):
     (value, parse_units) = tup[1:]
     if to_units is not None:
+        from rsciio.utils._units import _UREG
+
         v = value * _UREG(parse_units)
         value = float("%.6e" % v.to(to_units).magnitude)
     return value