+++ rsciio/tvips/_api.py
@@ -23,10 +23,8 @@ import warnings
 from datetime import datetime, timezone
 
 import dask
-import dask.array as da
 import numpy as np
 import pint
-from dask.diagnostics import ProgressBar
 from dateutil.parser import parse as dtparse
 
 from rsciio._docstrings import (
@@ -36,13 +34,10 @@ from rsciio._docstrings import (
     SHOW_PROGRESSBAR_DOC,
     SIGNAL_DOC,
 )
-from rsciio.utils.tools import (
-    _UREG,
-    DTBox,
-    dummy_context_manager,
-    jit_ifnumba,
-    sarray2dict,
-)
+from rsciio.utils._array import sarray2dict
+from rsciio.utils._context_manager import get_progress_bar_context_manager
+from rsciio.utils._decorator import jit_ifnumba
+from rsciio.utils._dictionary import DTBox
 
 _logger = logging.getLogger(__name__)
 
@@ -95,6 +90,8 @@ def _guess_image_mode(signal):
     unit = signal["axes"][-2]["units"]
     mode = None
     try:
+        from rsciio.utils._units import _UREG
+
         pixel_size = scale * _UREG(unit)
     except (AttributeError, pint.UndefinedUnitError):
         pass
@@ -127,6 +124,8 @@ def _get_main_header_from_signal(signal, version=2, frame_header_extra_bytes=0):
     else:
         to_unit = ""
     if to_unit:
+        from rsciio.utils._units import _UREG
+
         scale = round((scale * _UREG(unit)).to(to_unit).magnitude)
         offsetx = round((offsetx * _UREG(unit)).to(to_unit).magnitude)
         offsety = round((offsety * _UREG(unit)).to(to_unit).magnitude)
@@ -337,6 +336,8 @@ def file_reader(
         all_metadata.append(records[metadata_keys])
         all_array_data.append(records["data"])
     if lazy:
+        import dask.array as da
+
         data_stack = da.concatenate(all_array_data, axis=0)
     else:
         data_stack = np.concatenate(all_array_data, axis=0)
@@ -578,6 +579,8 @@ def file_writer(
         nav_units = signal["axes"][-3]["units"]
         nav_increment = signal["axes"][-3]["scale"]
         try:
+            from rsciio.utils._units import _UREG
+
             time_increment = (nav_increment * _UREG(nav_units)).to("ms").magnitude
         except (AttributeError, pint.UndefinedUnitError, pint.DimensionalityError):
             time_increment = 1
@@ -637,8 +640,7 @@ def file_writer(
         file_memmap["rotidx"] = rotator + 1
         data = fdata[current_frame : current_frame + frames_saved]
         if signal["attributes"]["_lazy"]:
-            cm = ProgressBar if show_progressbar else dummy_context_manager
-            with cm():
+            with get_progress_bar_context_manager(show_progressbar)():
                 data.store(file_memmap["data"])
         else:
             file_memmap["data"] = data