+++ rsciio/utils/_array.py
@@ -1,0 +1,84 @@
+# -*- coding: utf-8 -*-
+# Copyright 2007-2025 The HyperSpy developers
+#
+# This file is part of RosettaSciIO.
+#
+# RosettaSciIO is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# RosettaSciIO is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with RosettaSciIO. If not, see <https://www.gnu.org/licenses/#GPL>.
+
+from collections import OrderedDict
+
+import numpy as np
+
+
+def sarray2dict(sarray, dictionary=None):
+    """Converts a struct array to an ordered dictionary
+
+    Parameters
+    ----------
+    sarray: struct array
+    dictionary: None or dict
+        If dictionary is not None the content of sarray will be appended to the
+        given dictonary
+
+    Returns
+    -------
+    Ordered dictionary
+
+    """
+    if dictionary is None:
+        dictionary = OrderedDict()
+    for name in sarray.dtype.names:
+        dictionary[name] = sarray[name][0] if len(sarray[name]) == 1 else sarray[name]
+    return dictionary
+
+
+def dict2sarray(dictionary, sarray=None, dtype=None):
+    """Populates a struct array from a dictionary
+
+    Parameters
+    ----------
+    dictionary: dict
+    sarray: struct array or None
+        Either sarray or dtype must be given. If sarray is given, it is
+        populated from the dictionary.
+    dtype: None, numpy dtype or dtype list
+        If sarray is None, dtype must be given. If so, a new struct array
+        is created according to the dtype, which is then populated.
+
+    Returns
+    -------
+    Structure array
+
+    """
+    if sarray is None:
+        if dtype is None:
+            raise ValueError("Either sarray or dtype need to be specified.")
+        sarray = np.zeros((1,), dtype=dtype)
+    for name in set(sarray.dtype.names).intersection(set(dictionary.keys())):
+        if len(sarray[name]) == 1:
+            sarray[name][0] = dictionary[name]
+        else:
+            sarray[name] = dictionary[name]
+    return sarray
+
+
+def is_dask_array(x):
+    """Check if x is a dask array
+
+    Parameters
+    ----------
+    x : object
+        Object to be checked.
+    """
+    return hasattr(x, "__dask_keys__")